+++ src/containers/HeaderContainer/HeaderContainer.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="header">
     <!-- Modals ***************************************** -->
-    <disconnected-modal />
+    <disconnected-modal ref="mewConnectDisconnected" />
     <settings-modal
       v-if="wallet !== null"
       ref="settings"
@@ -195,6 +195,11 @@ import BigNumber from 'bignumber.js';
 import MobileMenu from './components/MobileMenu';
 import DisconnectedModal from '@/components/DisconnectedModal';
 
+const events = {
+  issueModal: 'issueModal',
+  mewConnectDisconnected: 'mewConnectDisconnected'
+};
+
 export default {
   components: {
     blockie: Blockie,
@@ -286,6 +291,7 @@ export default {
         .catch(e => {
           ErrorHandler(e, false);
         });
+      // this.disconnectMewConnectModal();
     }
   },
   mounted() {
@@ -302,12 +308,6 @@ export default {
       store.set('locale', storedLocale.langCode);
       this.currentFlag = storedLocale.flag;
     }
-
-    // https://github.com/MyEtherWallet/MyEtherWallet/projects/2#card-12172489
-    // trivial statement to convert dialects to primary language tags, with the exception of Chinese
-    // if (!/zh[-_]/.test(this.currentFlag)) {
-    //   this.currentFlag = this.currentFlag.split(/[-_]/)[0];
-    // }
 
     this.currentName = this.supportedLanguages.find(
       item => item.flag === this.currentFlag
@@ -333,9 +333,22 @@ export default {
         this.resolver = resolve;
       }
     });
+
+    // this.disconnectMewConnectModal();
+
+    this.$eventHub.$on('mewConnectDisconnected', () => {
+      this.isMobileMenuOpen = false;
+      this.$refs.mewConnectDisconnected.$refs.disconnected.show();
+      this.$refs.mewConnectDisconnected.$refs.disconnected.$on('hidden', () => {
+        this.$router.push('/access-my-wallet');
+      });
+    });
   },
   beforeDestroy() {
-    this.$eventHub.$off('issueModal');
+    Object.values(events).forEach(evt => {
+      this.$eventHub.$off(evt);
+    });
+    // this.$eventHub.$off('issueModal');
   },
   created() {
     try {