+++ src/store/actions.js
@@ -87,6 +87,14 @@ const createAndSignTx = function({ commit }, val) {
 
 const decryptWallet = function({ commit, dispatch }, params) {
   // params[0] = wallet, params[1] = provider
+  if (params[0].identifier === MEW_CONNECT) {
+    params[0].mewConnect().on('RtcClosedEvent', () => {
+      if (params[0].mewConnect().getConnectonState()) {
+        this._vm.$eventHub.$emit('mewConnectDisconnected');
+        dispatch('clearWallet');
+      }
+    });
+  }
   commit('DECRYPT_WALLET', params[0]);
   dispatch('setWeb3Instance', params[1]);
 };